Fix placeholder showing up in popup instead of output
Fixes #55

Directly assigning the image source didn't update is onClick
event handler which dynamically assigns the image src before toggling
the pop. Might be good to change that in the future maybe but in the
meantime this means people can click on the output and see the image,
not the placeholder.